# api/openapi.yaml
openapi: 3.1.0
info:
  title: Worthly API
  version: 4.3.0
  description: |
    Canonical MVP API contract for Worthly.

    Protected resources are Bearer-token OAuth2 Resource Server endpoints.
    The Next.js `web_session` cookie is never an API credential.

    Aggregate amounts are always grouped by ISO 4217 currency. No schema
    exposes a single cross-currency Money total.
servers:
  - url: https://worthly.example.com/api/v1
security:
  - oauth2: [worthly.read]
...
paths:
  /register:
    post:
      tags: [Identity]
      operationId: register
      security: []
      description: |
        Creates an isolated user. Does not issue tokens. The caller then
        signs in through OAuth. Duplicate email is 409 `email_taken`.
        Password minimum length is 8 characters.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RegisterRequest'}
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Owner'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
  /me:
    get:
      tags: [Identity]
      operationId: getMe
      responses:
        '200':
          description: Current owner
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Owner'}
        '401': {$ref: '#/components/responses/Unauthorized'}
    patch:
      tags: [Identity]
      operationId: updateMe
      security: [{oauth2: [worthly.write]}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OwnerPatch'}
      responses:
        '200':
          description: Updated owner preferences
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Owner'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
    delete:
      tags: [Identity]
      operationId: deleteMe
      security: [{oauth2: [worthly.write]}]
      description: |
        Irreversibly deletes the authenticated user after `confirm=true`.
        Disconnects and purges every connection, removes sessions and tokens,
        then deletes `app_user`. Audit rows stay with `user_id` nulled.
        The BFF MUST clear the `web_session` cookie afterwards.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DeleteAccountRequest'}
      responses:
        '204': {description: Account deleted}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /me/logout:
    post:
      tags: [Identity]
      operationId: logoutCurrent
      description: |
        Revokes the current access token's refresh-token family and any
        associated web or device session. The BFF MUST also clear the
        `web_session` cookie. Mobile MUST delete secure-storage material.
      responses:
        '204': {description: Logged out}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
  /connections/banks:
    get:
      tags: [Connections]
      operationId: listAvailableBanks
      parameters:
        - name: country
          in: query
          schema: {type: string, default: PT, minLength: 2, maxLength: 2}
      responses:
        '200':
          description: Dynamically discovered AIS banks
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/BankChoice'}
  /connections/trading-212:
    post:
      tags: [Connections]
      operationId: connectTrading212
      security: [{oauth2: [worthly.write]}]
      description: |
        Stores a read-only Trading 212 API key and secret encrypted at rest
        for the current user. Secrets never appear in responses, logs or
        audit metadata. Validates the key against Trading 212 when reachable.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Trading212ConnectRequest'}
      responses:
        '200':
          description: Connection upserted
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Connection'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /connections/enable-banking/authorize:
    post:
      tags: [Connections]
      operationId: startEnableBankingAuthorization
      security: [{oauth2: [worthly.write]}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BankAuthorizationRequest'}
      responses:
        '200':
          description: Authorization redirect
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthorizationRedirect'}
        '400': {$ref: '#/components/responses/BadRequest'}
  /connections/enable-banking/callback:
    get:
      tags: [Connections]
      operationId: enableBankingCallback
      security: []
      parameters:
        - {name: code, in: query, schema: {type: string}}
        - {name: state, in: query, required: true, schema: {type: string}}
        - {name: error, in: query, schema: {type: string}}
        - {name: error_description, in: query, schema: {type: string}}
      responses:
        '302':
          description: Redirect to allowlisted Worthly client result page
        '400': {$ref: '#/components/responses/BadRequest'}
  /connections/{connectionId}/sync:
    post:
      tags: [Connections]
      operationId: requestSync
      security: [{oauth2: [worthly.write]}]
      parameters:
        - {$ref: '#/components/parameters/ConnectionId'}
      responses:
        '202':
          description: Sync accepted
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SyncRun'}
        '409':
          description: Sync already running
          content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
  /connections/{connectionId}/sync-runs:
    get:
      tags: [Connections]
      operationId: listSyncRuns
      parameters:
        - {$ref: '#/components/parameters/ConnectionId'}
        - {name: page, in: query, schema: {type: integer, minimum: 0, default: 0}}
        - {name: size, in: query, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
      responses:
        '200':
          description: Sync runs newest first
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SyncRunPage'}
        '404': {$ref: '#/components/responses/NotFound'}
  /connections/{connectionId}:
    delete:
      tags: [Connections]
      operationId: disconnect
      description: |
        Sets status DISABLED and stops synchronization. Historical
        normalized data is retained until purge.
      security: [{oauth2: [worthly.write]}]
      parameters:
        - {$ref: '#/components/parameters/ConnectionId'}
      responses:
        '204': {description: Disconnected}
  /connections/{connectionId}/purge:
    post:
      tags: [Connections]
      operationId: purgeConnectionData
      description: |
        Deletes imported provider data for this connection after explicit
        confirmation. Audit events are retained. See domain data lifecycle.
      security: [{oauth2: [worthly.write]}]
      parameters:
        - {$ref: '#/components/parameters/ConnectionId'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PurgeRequest'}
      responses:
        '204': {description: Purged}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://worthly.example.com/oauth2/authorize
          tokenUrl: https://worthly.example.com/oauth2/token
          refreshUrl: https://worthly.example.com/oauth2/token
          scopes:
            openid: OpenID identity
            profile: Basic owner profile
            worthly.read: Read Worthly financial data
            worthly.write: Modify Worthly metadata/settings
  parameters:
...
    ConnectionId: {name: connectionId, in: path, required: true, schema: {type: string, format: uuid}}
...
  responses:
    Unauthorized:
      description: Unauthorized
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
    NotFound:
      description: Not found
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
    BadRequest:
      description: Bad request
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
    Conflict:
      description: Conflict
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
    Forbidden:
      description: Forbidden
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
  schemas:
...
    Owner:
      type: object
      required: [id, email, reportingTimezone, reportingCurrency]
      properties:
        id: {type: string, format: uuid}
        email: {type: string, format: email}
        reportingTimezone: {type: string, example: Europe/Lisbon}
        reportingCurrency: {type: string, example: EUR}
    OwnerPatch:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        reportingTimezone:
          type: string
          minLength: 1
          maxLength: 64
          description: IANA timezone identifier validated by the server, e.g. Europe/Lisbon.
        reportingCurrency:
          type: string
          pattern: '^[A-Z]{3}$'
          description: ISO 4217 reporting currency preference. Does not imply FX conversion.
    RegisterRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email: {type: string, format: email}
        password: {type: string, minLength: 8, maxLength: 128}
        reportingTimezone:
          type: string
          minLength: 1
          maxLength: 64
        reportingCurrency:
          type: string
          pattern: '^[A-Z]{3}$'
        invite:
          type: string
          maxLength: 128
          description: Required when the API is configured with an invite code.
    DeleteAccountRequest:
      type: object
      required: [confirm]
      additionalProperties: false
      properties:
        confirm:
          type: boolean
          description: Must be true. Any other value is 400 `confirm_required`.
...
    ConnectionStatus:
      type: string
      enum: [ACTIVE, REAUTH_REQUIRED, CONFIGURATION_REQUIRED, ERROR, DISABLED]
    Connection:
      type: object
      required: [id, provider, status]
      properties:
        id: {type: string, format: uuid}
        provider: {type: string, enum: [ENABLE_BANKING, TRADING_212]}
        status: {$ref: '#/components/schemas/ConnectionStatus'}
        institutionName: {type: string}
        institutionCountry: {type: string}
        lastSuccessfulSyncAt: {type: string, format: date-time}
        consentExpiresAt: {type: string, format: date-time}
        lastErrorCode: {type: string}
    BankChoice:
      type: object
      required: [name, country]
      properties:
        name: {type: string}
        country: {type: string}
        logoUrl: {type: string, format: uri}
    BankAuthorizationRequest:
      type: object
      required: [name, country, returnClient]
      properties:
        name: {type: string}
        country: {type: string, minLength: 2, maxLength: 2}
        returnClient: {type: string, enum: [WEB, MOBILE]}
    Trading212ConnectRequest:
      type: object
      required: [apiKey, apiSecret]
      additionalProperties: false
      properties:
        apiKey: {type: string, minLength: 1, maxLength: 256}
        apiSecret: {type: string, minLength: 1, maxLength: 256}
        environment:
          type: string
          enum: [LIVE, DEMO]
          description: LIVE is the real account; DEMO is Practice. If omitted, LIVE is tried first then DEMO.
    AuthorizationRedirect:
      type: object
      required: [url, expiresAt]
      properties:
        url: {type: string, format: uri}
        expiresAt: {type: string, format: date-time}
    PurgeRequest:
      type: object
      required: [confirm]
      properties:
        confirm:
          type: boolean
          description: Must be true. False or omitted is 400.
...
    SyncRun:
      type: object
      required: [id, status, startedAt]
      properties:
        id: {type: string, format: uuid}
        status: {type: string, enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, RATE_LIMITED]}
        startedAt: {type: string, format: date-time}
        finishedAt: {type: string, format: date-time}
        importedCount: {type: integer}
        updatedCount: {type: integer}
        nextRetryAt: {type: string, format: date-time}
        errorCode: {type: string}
    SyncRunPage:
      type: object
      required: [items, page, size, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/SyncRun'}}
        page: {type: integer}
        size: {type: integer}
        total: {type: integer, format: int64}
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string, format: uri-reference}
        correlationId: {type: string}